(refac) extract formfields as generic components
